+++ src/Fonts/Generic.ts
@@ -1,31 +1,42 @@
 import { FontAssetType } from 'fantasticon';
-import { workspace, Uri } from 'vscode';
+import { workspace, Uri, Progress } from 'vscode';
 import path from 'path';
 import fs from 'fs';
 
 import Handlebars from 'handlebars';
+import { IconFontBundlerFontConfig, IconFontBundlerItem } from '../Types/Types';
+import Log from '../Utils/Log';
 
 export default {
-  async generateAssets(oFontOptions, oFontConfig) {
-    oFontOptions.assetTypes.forEach(async (sExtension) => {
-      let sTemplatePath = oFontOptions.templates[sExtension];
+  async generateAssets(
+    oFontOptions: IconFontBundlerFontConfig,
+    oFontConfig: IconFontBundlerItem,
+    progress?: Progress<any>,
+    multiplier = 1
+  ) {
+    const message = Log.fonts(`Generating assets...`);
+    progress?.report({ increment: 10 * multiplier, message });
+
+    const templates: any = oFontOptions.templates || {};
+    oFontOptions.assetTypes?.forEach(async (sExtension) => {
+      const sTemplatePath = templates[sExtension];
       if (sTemplatePath) {
-        let sTemplate;
-        sTemplate = fs.readFileSync(sTemplatePath, { encoding: 'utf8' });
+        const sTemplate = fs.readFileSync(sTemplatePath, { encoding: 'utf8' });
 
         try {
-          let sFile = Handlebars.compile(sTemplate)(oFontOptions, {
+          const sFile = Handlebars.compile(sTemplate)(oFontOptions, {
             helpers: {
+              // @ts-ignore
               fontSrc: this.renderSrcAttribute(oFontOptions),
-              codepoint(codepoint) {
+              codepoint(codepoint: any) {
                 return codepoint.toString(16);
               },
             },
           });
-          let oCodePointsUri = Uri.file(path.join(oFontOptions.outputDir, `${oFontOptions.name}.${sExtension}`));
+          const oCodePointsUri = Uri.file(path.join(oFontOptions.outputDir, `${oFontOptions.name}.${sExtension}`));
           workspace.fs.writeFile(oCodePointsUri, Buffer.from(sFile));
         } catch (oError) {
-          debugger;
+          //debugger;
         }
       }
     });
@@ -36,19 +47,22 @@ export default {
       formatValue: 'embedded-opentype',
       getSuffix: () => '#iefix',
     },
-    [FontAssetType.WOFF2]: { formatValue: 'woff2' },
-    [FontAssetType.WOFF]: { formatValue: 'woff' },
-    [FontAssetType.TTF]: { formatValue: 'truetype' },
-    [FontAssetType.SVG]: { formatValue: 'svg', getSuffix: (name) => `#${name}` },
+    [FontAssetType.WOFF2]: { formatValue: 'woff2', getSuffix: null },
+    [FontAssetType.WOFF]: { formatValue: 'woff', getSuffix: null },
+    [FontAssetType.TTF]: { formatValue: 'truetype', getSuffix: null },
+    [FontAssetType.SVG]: { formatValue: 'svg', getSuffix: (name: string | undefined) => `#${name}` },
   },
 
-  renderSrcAttribute({ name, fontTypes, fontsUrl }) {
-    return fontTypes
-      .map((fontType) => {
-        const { formatValue, getSuffix } = this.renderSrcOptions[fontType];
-        const suffix = getSuffix ? getSuffix(name) : '';
-        return `url("${fontsUrl || '.'}/${name}.${fontType}?${suffix}") format("${formatValue}")`;
-      })
-      .join(',\n');
+  renderSrcAttribute(oFontOptions: IconFontBundlerFontConfig) {
+    const { name, fontTypes, fontsUrl } = oFontOptions;
+    return (
+      fontTypes
+        ?.map((fontType) => {
+          const { formatValue, getSuffix } = this.renderSrcOptions[fontType];
+          const suffix = getSuffix ? getSuffix(name) : '';
+          return `url("${fontsUrl || '.'}/${name}.${fontType}?${suffix}") format("${formatValue}")`;
+        })
+        .join(',\n') || ''
+    );
   },
 };