+++ src/Fonts/types/index.ts
@@ -1,9 +1,7 @@
 import Fantasticon from './Fantasticon';
 import FontAwesome from './FontAwesome';
-import Material from './Material';
 
 export default {
-    fantasticon: Fantasticon,
-    fontawesome: FontAwesome,
-    material: Material
-}
+  fantasticon: Fantasticon,
+  fontawesome: FontAwesome,
+};