+++ src/test/suite/extension.test.ts
@@ -1,0 +1,13 @@
+import assert from 'assert';
+import Utils from '../../Utils/Utils';
+
+suite('Startup extension tests', () => {
+  test('Extension is registered', () => {
+    assert(!!Utils.getExtensionInfo());
+  });
+
+  test('Extension is active', () => {
+    const extension = Utils.getExtensionInfo();
+    assert(extension && extension.isActive);
+  });
+});