+++ src/test/suite/fontawesome.test.ts
@@ -1,0 +1,6 @@
+import assert from 'assert';
+import Utils from '../../Utils/Utils';
+
+suite('Fontawesome tests', () => {
+  test('Extension is registered', () => {});
+});