+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 ## 0.0.1 (26-07-2021)
 
-- Initial version of the icon-font-bundler extension with fontasticon and fontawesome
+- Initial version of the icon-font-bundler extension with fantasticon and fontawesome